+++ internal/user/DBstorage.go
@@ -56,3 +56,22 @@ func (s *PostgresStorage) Exists(email string) bool {
 	}
 	return exists
 }
+
+func (s *PostgresStorage) GetUserById(id string) (User, bool) {
+	var user User
+	err := s.db.QueryRow("SELECT id, email, password, role FROM users WHERE id=$1", id).Scan(&user.ID, &user.Email, &user.Password, &user.Role, &user.Name, &user.Lastname)
+	if err != nil {
+		log.Fatal().Err(err).Msg("Failed to fetch user ID")
+		return User{}, false
+	}
+	return user, true // TODO do we need pinter here
+}
+
+func (s *PostgresStorage) UpdateUser(id string, user User) bool {
+	_, err := s.db.Exec("UPDATE users SET email=$1, password=$2, role=$3, name=$4, lastname=$5 WHERE id=$6", user.Email, user.Password, user.Role, user.Name, user.Lastname, user.ID)
+	if err != nil {
+		log.Fatal().Err(err).Msg("Failed to update user")
+		return false
+	}
+	return true
+}
