+++ internal/user/handler.go
@@ -1,33 +1,77 @@
-package user
+package User
 
 import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+
+	"github.com/rs/zerolog/log"
 )
 
-type UserResource struct {
-	S *Storage
+// var s InMemStorage
+
+type CreateUserRequestBody struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
 }
 
-func (p *UserResource) GetAllUsers(w http.ResponseWriter, r *http.Request) {
-	users := p.S.GetAllUsers()
-	res := map[string][]User{"users": users}
+type service interface {
+	SignUp(email, password string) error
+	GetUsers() []User
+}
 
-	err := json.NewEncoder(w).Encode(res)
+type Handler struct {
+	s service
+}
 
-	if err != nil {
-		fmt.Println("Failed to encode: ", err.Error())
-		w.WriteHeader(http.StatusInternalServerError)
-
-		return
+func NewHandler(s service) Handler {
+	return Handler{
+		s: s,
 	}
 }
 
-func (p *UserResource) GetUserById(w http.ResponseWriter, r *http.Request) {
+func (h Handler) Create(w http.ResponseWriter, r *http.Request) {
+	var reqBody CreateUserRequestBody
+
+	err := json.NewDecoder(r.Body).Decode(&reqBody)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		log.Debug().Err(err).Msg("Failed to decode JSON response")
+		return
+	}
+
+	err = h.s.SignUp(reqBody.Email, reqBody.Password)
+	if err != nil {
+		if err.Error() == "user already exists" {
+			w.WriteHeader(http.StatusConflict)
+			json.NewEncoder(w).Encode(map[string]string{"error": "User already exists"})
+		} else {
+			w.WriteHeader(http.StatusInternalServerError)
+			json.NewEncoder(w).Encode(map[string]string{"error": "Internal server error"})
+		}
+		return
+	}
+
+	w.WriteHeader(http.StatusCreated)
+}
+
+// Створила цю функц щоб перевірити чи відправляються юзери на сервер
+func (h Handler) GetUsers(w http.ResponseWriter, r *http.Request) {
+	users := h.s.GetUsers()
+	err := json.NewEncoder(w).Encode(users)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		log.Debug().Err(err).Msg("Failed to encode JSON response")
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+}
+
+func (h Handler) GetUserById(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 
-	user, ok := p.S.GetUserById(id)
+	user, ok := h.s.GetUserById(id)
 
 	if !ok {
 		w.WriteHeader(http.StatusBadRequest)
@@ -35,36 +79,14 @@ func (p *UserResource) GetUserById(w http.ResponseWriter, r *http.Request) {
 	}
 
 	err := json.NewEncoder(w).Encode(user)
-
 	if err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
+	w.Header().Set("Content-Type", "application/json")
 }
 
-func (p *UserResource) CreateUser(w http.ResponseWriter, r *http.Request) {
-	var reqBody User
-
-	userId, ok := p.S.CreateUser(reqBody)
-
-	if !ok {
-		fmt.Print("Error creating user")
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	reqBody.ID = userId
-
-	err := json.NewDecoder(r.Body).Decode(&reqBody)
-
-	if err != nil {
-		fmt.Println("Failed to decode: ", err.Error())
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-}
-
-func (p *UserResource) UpdateUser(w http.ResponseWriter, r *http.Request) {
+func (h Handler) UpdateUser(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 
 	var reqBody User
@@ -77,11 +99,12 @@ func (p *UserResource) UpdateUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	ok := p.S.UpdateUser(id, reqBody)
+	err = h.s.UpdateUser(reqBody, id)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			json.NewEncoder(w).Encode(map[string]string{"error": "Failed to fetch user"})
+		}
 
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
+	w.Header().Set("Content-Type", "application/json")
 
 }
